% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AK_marine_area.R
\name{AK_marine_area}
\alias{AK_marine_area}
\title{AK_marine_area}
\usage{
AK_marine_area(area = "NMFS Reporting Area", prj = "dd")
}
\arguments{
\item{area}{the area of interest.
"NMFS Reporting Area" (Default)
"Ecosystem Area"
"Ecosystem Subarea"
"Ecosystem Area Inside" (GOA only)
"Ecosystem Subarea Inside (GOA only)
"ADFG Stat Area"
"BSIERP Region"
"BSIERP Name"}

\item{prj}{Projection: Default is dd, which sets WGS 1984, ak_prj sets NAD 1983 Alaska Albers}
}
\value{

}
\description{
This package imports sf objects from RDS files
containing spatial boundaries of marine management areas.
Areas include NMFS areas, ADFG stat areas, Ecosystem boundaries,
and BSIERP regions. WGS 1984 and an Alaska Albers projection versions
are available. Data were originally compiled in a filegeodatabase
produced by Brett Holycross at PSMFC.
}
\examples{
AK_marine_area(area="NMFS Reporting Area", prj="dd")
}
